+++ examples/resources/scalr_drift_detection/resource.tf
@@ -1,0 +1,4 @@
+resource "scalr_drift_detection" "example" {
+  environment_id = "env-xxxxx"
+  check_period   = "weekly"
+}