+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+### Added
+
+- **New resource:** `scalr_drift_detection` ([#473](https://github.com/Scalr/terraform-provider-scalr/pull/473))
+
+
 ## [3.10.0] - 2025-10-23
 
 ### Changed